move PGN tree view to dedicated file
