+++ lib/src/view/analysis/analysis_board.dart
@@ -15,7 +15,7 @@ import 'package:lichess_mobile/src/model/common/chess.dart';
 import 'package:lichess_mobile/src/model/common/eval.dart';
 import 'package:lichess_mobile/src/model/engine/evaluation_service.dart';
 import 'package:lichess_mobile/src/model/settings/board_preferences.dart';
-import 'package:lichess_mobile/src/view/analysis/annotations.dart';
+import 'package:lichess_mobile/src/view/pgn/annotations.dart';
 
 class AnalysisBoard extends ConsumerStatefulWidget {
   const AnalysisBoard(
